// src/app/model/account.model.ts
export class AccountModel {
  id: "";
  userId: "";
  userName: "";
  password: "";
  createdDate:any = null;
  modifiedDate:any = null
  createdBy: "";
  modifiedBy: "";
  oldPassword: "";
  roleId: "";
  status: 0
  firstName: "";
  lastName: "";
  email: "";
  address: "";
  phone: "";
  identityNo: "";
  className: "";
  birthday: ""; 
  avatar: "";
  roleName: "";
  roleDescription: "";
}